Add scheduled task runner.
