+++ README.md
@@ -8,3 +8,5 @@ using user as user name and pass as password for security authorization requires
 using swagger to check API detailse as follow: http://localhost:9080/swagger-ui.html
 
 using h2 web console: http://localhost:9080/h2-console/ which have default user name and password.
+
+using spring boot actutor to check source code information: http://localhost:9080/actuator/info
