+++ src/main/java/com/impulse/config/WebSecurityConfig.java
@@ -1,0 +1,25 @@
+package com.impulse.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.h2.H2ConsoleProperties;
+import org.springframework.boot.autoconfigure.security.SecurityProperties;
+import org.springframework.core.annotation.Order;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@EnableWebSecurity
+@Order(SecurityProperties.BASIC_AUTH_ORDER - 10)
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+  @Autowired
+  private H2ConsoleProperties console;
+
+  @Override
+  public void configure(final WebSecurity web) throws Exception {
+    final String path = this.console.getPath();
+    web.ignoring().antMatchers("/v2/api-docs", "/swagger-resources/**", "/swagger-ui.html**",
+        "/webjars/**", "/configuration/**", path + "/**", "/actuator/**");
+    super.configure(web);
+  }
+
+}
