+++ src/main/java/com/impulse/task/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * this package is for scheduled task including batch task.
+ */
+package com.impulse.task;
