+++ src/main/resources/application.properties
@@ -23,3 +23,8 @@ management.info.git.mode=full
 
 # PROJECT INFORMATION (ProjectInfoProperties)
 spring.info.git.location=classpath:git.properties
+
+# TASK SCHEDULING  (TaskSchedulingProperties)
+spring.task.scheduling.pool.size=1
+spring.task.scheduling.thread-name-prefix=impulse-scheduling-
+
